# quizano/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quizano Simple App API
  version: 1.0.0
  description: Node.js API for a simplified Quizano app with student and admin flows.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/exams:
    get:
      tags: [Exams]
      summary: List exams
      description: Students only receive active exams.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
            enum: [free, scheduled]
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: availableNow
          schema:
            type: boolean
      responses:
        '200':
          description: Exams list with metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExamWithMeta'
    post:
      tags: [Exams]
      summary: Create exam
      description: Admin only. Supports nested questions creation.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExamRequest'
      responses:
        '201':
          description: Created exam
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      exam:
                        $ref: '#/components/schemas/Exam'
                      questionsCount:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/exams/{examId}:
    get:
      tags: [Exams]
      summary: Get exam by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Exam detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ExamWithMeta'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Exams]
      summary: Update exam
      description: Admin only. Can include questions array for full replacement.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExamRequest'
      responses:
        '200':
          description: Updated exam
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Exam'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Exams]
      summary: Delete exam and linked questions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                      examId:
                        type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/exams/{examId}/questions:
    get:
      tags: [Exams]
      summary: Get questions by exam
      description: Students receive no correctOptionId/explanation. Admin can request includeAnswers=true.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
        - in: query
          name: includeAnswers
          schema:
            type: boolean
      responses:
        '200':
          description: Questions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/QuestionPublic'
                        - $ref: '#/components/schemas/Question'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Exams]
      summary: Replace all questions of exam
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questions]
              properties:
                questions:
                  type: array
                  items:
                    $ref: '#/components/schemas/QuestionInput'
      responses:
        '200':
          description: Questions replaced
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      examId:
                        type: string
                      questionsCount:
                        type: integer
  /api/v1/exams/{examId}/questions/import:
    post:
      tags: [Exams]
      summary: Import questions from row payload
      description: Admin only. Import rows similar to excel parsed output.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportQuestionsRequest'
      responses:
        '200':
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      importedCount:
                        type: integer
                      examId:
                        type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Validation failed
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Resource not found
    Conflict:
      description: Conflict state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Username already exists
  schemas:
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            details:
              nullable: true
...
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [free, scheduled]
        duration:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        startTime:
          type: string
          format: date-time
          nullable: true
        endTime:
          type: string
          format: date-time
          nullable: true
    ExamWithMeta:
      allOf:
        - $ref: '#/components/schemas/Exam'
        - type: object
          properties:
            questionCount:
              type: integer
            availableNow:
              type: boolean
    Option:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
    QuestionInput:
      type: object
      required: [content, options, correctOptionId]
      properties:
        id:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        correctOptionId:
          type: string
        explanation:
          type: string
    Question:
      allOf:
        - type: object
          properties:
            examId:
              type: string
        - $ref: '#/components/schemas/QuestionInput'
    QuestionPublic:
      type: object
      properties:
        id:
          type: string
        examId:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
    CreateExamRequest:
      type: object
      required: [title, description, type, duration, status]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [free, scheduled]
        duration:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionInput'
    UpdateExamRequest:
      allOf:
        - $ref: '#/components/schemas/CreateExamRequest'
    ImportQuestionRow:
      type: object
      properties:
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
        explanation:
          type: string
    ImportQuestionsRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ImportQuestionRow'
...